+++ lib/smurff-cpp/SmurffCpp/DataMatrices/SparseMatrixData.cpp
@@ -2,45 +2,45 @@
 
 using namespace smurff;
 
-SparseMatrixData::SparseMatrixData(Eigen::SparseMatrix<double> Y)
-   : FullMatrixData<Eigen::SparseMatrix<double>>(Y)
+SparseMatrixData::SparseMatrixData(Eigen::SparseMatrix<float> Y)
+   : FullMatrixData<Eigen::SparseMatrix<float>>(Y)
 {
    this->name = "SparseMatrixData [fully known]";
 }
 
-void SparseMatrixData::getMuLambda(const SubModel& model, uint32_t mode, int d, Eigen::VectorXd& rr, Eigen::MatrixXd& MM) const
+void SparseMatrixData::getMuLambda(const SubModel& model, uint32_t mode, int d, Eigen::VectorXf& rr, Eigen::MatrixXf& MM) const
 {
     const auto& Y = this->Y(mode);
     auto Vf = *model.CVbegin(mode);
     auto &ns = noise();
 
-    for (Eigen::SparseMatrix<double>::InnerIterator it(Y, d); it; ++it) 
+    for (Eigen::SparseMatrix<float>::InnerIterator it(Y, d); it; ++it) 
     {
         const auto &col = Vf.col(it.row());
         auto p = pos(mode, d, it.row());
-        double noisy_val = ns.sample(model, p, it.value());
+        float noisy_val = ns.sample(model, p, it.value());
         rr.noalias() += col * noisy_val; // rr = rr + (V[m] * y[d]) * alpha
     }
 
     MM.noalias() += ns.getAlpha() * VV[mode]; // MM = MM + VV[m]
 }
 
-double SparseMatrixData::train_rmse(const SubModel& model) const
+float SparseMatrixData::train_rmse(const SubModel& model) const
 {
    return std::sqrt(sumsq(model) / this->size());
 }
 
-double SparseMatrixData::var_total() const
+float SparseMatrixData::var_total() const
 {
-   const double cwise_mean = this->sum() / this->size();
-   const double cwise_mean_squared = std::pow(cwise_mean, 2);
-   double se = 0.0;
+   const float cwise_mean = this->sum() / this->size();
+   const float cwise_mean_squared = std::pow(cwise_mean, 2);
+   float se = 0.0;
 
    #pragma omp parallel for schedule(guided) reduction(+:se)
    for(int c = 0; c < Y().cols(); ++c)
    {
       int r = 0;
-      for (Eigen::SparseMatrix<double>::InnerIterator it(Y(), c); it; ++it)
+      for (Eigen::SparseMatrix<float>::InnerIterator it(Y(), c); it; ++it)
       {
          se += (r - it.row()) * cwise_mean_squared; // handle implicit zeroes
          se += std::pow(it.value() - cwise_mean, 2);
@@ -50,7 +50,7 @@ double SparseMatrixData::var_total() const
       se += (r - Y().rows()) * cwise_mean_squared; // handle implicit zeroes
    }
 
-   double var = se / this->size();
+   float var = se / this->size();
    if (var <= 0.0 || std::isnan(var))
    {
       // if var cannot be computed using 1.0
@@ -60,15 +60,15 @@ double SparseMatrixData::var_total() const
    return var;
 }
 
-double SparseMatrixData::sumsq(const SubModel& model) const
+float SparseMatrixData::sumsq(const SubModel& model) const
 {
-   double sumsq = 0.0;
+   float sumsq = 0.0;
    
    #pragma omp parallel for schedule(guided) reduction(+:sumsq)
    for(int c = 0; c < Y().cols(); ++c)
    {
       int r = 0;
-      for (Eigen::SparseMatrix<double>::InnerIterator it(Y(), c); it; ++it)
+      for (Eigen::SparseMatrix<float>::InnerIterator it(Y(), c); it; ++it)
       {
          for(; r < it.row(); r++) //handle implicit zeroes
             sumsq += std::pow(model.predict({r, c}), 2);
