+++ lib/smurff-cpp/SmurffCpp/DataTensors/SparseMode.h
@@ -21,7 +21,7 @@ private:
 
    std::vector<std::uint64_t> m_row_ptr; // vector of offsets (in values and in indices) to each hyperplane
    MatrixXui32 m_indices; // [m_nnz x (m_nmodes - 1)] matrix of coordinates
-   std::vector<double> m_values; // vector of values
+   std::vector<float> m_values; // vector of values
    
 public:
    SparseMode();
@@ -30,7 +30,7 @@ public:
    // vals - vector of values
    // mode - index of dimension to fix
    // mode_size - size of dimension to fix
-   SparseMode(const MatrixXui32& idx, const std::vector<double>& vals, std::uint64_t mode, std::uint64_t mode_size);
+   SparseMode(const MatrixXui32& idx, const std::vector<float>& vals, std::uint64_t mode, std::uint64_t mode_size);
 
    std::uint64_t getNNZ() const;
 
@@ -38,7 +38,7 @@ public:
 
    std::uint64_t getNCoords() const;
 
-   const std::vector<double>& getValues() const;
+   const std::vector<float>& getValues() const;
 
    std::uint64_t getMode() const;
 
@@ -51,7 +51,7 @@ public:
    const MatrixXui32& getIndices() const;
 
 public:
-   std::pair<PVec<>, double> item(std::uint64_t hyperplane, std::uint64_t item) const;
+   std::pair<PVec<>, float> item(std::uint64_t hyperplane, std::uint64_t item) const;
    
    PVec<> pos(std::uint64_t hyperplane, std::uint64_t item) const;
 };
